"factored out watchdog; updates #1173"
