+++ software/controller/integration_tests/pinch_valve_test.h
@@ -12,6 +12,7 @@
 #include "hal.h"
 #include "pinch_valve.h"
 #include "system_timer.h"
+#include "watchdog.h"
 
 // test parameters
 static constexpr int MotorIndex{TEST_PARAM_1};
@@ -27,7 +28,7 @@ void RunTest() {
     pinch_valve.SetOutput(valve_open ? 1.0f : 0.0f);
     SystemTimer::singleton().Delay(Delay);
 
-    hal.WatchdogHandler();
+    Watchdog::pet();
 
     valve_open = !valve_open;
   }
