+++ software/controller/integration_tests/psol_test.h
@@ -10,6 +10,7 @@
 
 #include "hal.h"
 #include "system_timer.h"
+#include "watchdog.h"
 
 // test parameters
 static constexpr Duration Delay{milliseconds(10)};
@@ -27,7 +28,7 @@ void RunTest() {
     hal.psol_.PSolValue(psol_position);
     SystemTimer::singleton().Delay(Delay);
 
-    hal.WatchdogHandler();
+    Watchdog::pet();
 
     psol_position += step;
     if (psol_position >= PSolMax) {
